Fix logical error
was defining id in insert (3-to-4 request tables). Works fine if inserting into empty table, but if there's data it fails obviously. As part of this ended up moving "try again" from catch to `if(array_diff($shouldHave, $actuallyHave))` because when not inserted as above is not error so having "try-again" in "catch" wasn't doing anything.